Replace assets.go with _bindata
